Adds Core Map Config Files
Copied the very basics of what the bearcat and examples map have and threw together a unique one for the metro map, as well as including the new metro_define.dm file in baystation12.dme.

Please disregard the huge amount of changed lines in baystation12.dme, it was some automatic process which corrected the capitalization on many paths (which are actually valid now).